+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - added `createCleanTask` function to make it easier to clean builds.
 - changed CLI option `--colorful-log` to `--color`.
 - respect `NO_COLOR` environment variable (https://no-color.org/).
+- added `environment` argument to `runDartExe` function.
 
 ### Dartle Dart