+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## done, waiting for next release
 
+## 0.15.0
+
 ### Dartle Core
 
 - expose `runBasic` function to allow embedding Dartle in other CLI tools.